Create Osegbe-Nneamaka 2024714177
The python program implements a grading system using `if` and `elif` statements. It takes a student's score as input and print the corresponding grade based on the following scale.
-70_100: Grade A
-60_69: Grade B
-50_59: Grade C
-45_49: Grade D
-40_44: Grade E
-20_39: Grade F

For each valid score, the program also prints "YES"  to confirm the grading was successful. If the score is outside the valid range, it prints invalid score and "NO".

Created by: Osegbo Nneamaka Prisca 
Reg . No: 2024714177.